backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
google-generativeai>=0.8.0
python-decouple>=3.8
requests>=2.31.0
